[ISS] Added support for atomics

+++ engine/include/vp/gdbserver/gdbserver_engine.hpp
@@ -67,7 +67,7 @@ namespace vp
         static const int SIGNAL_STOP = 17;
 
         virtual int register_core(Gdbserver_core *core) = 0;
-        virtual void signal(Gdbserver_core *core, int signal, std::string reason="", uint64_t info=0) = 0;
+        virtual void signal(Gdbserver_core *core, int signal, std::string reason="", int info=0) = 0;
 
         virtual void lock() = 0;
         virtual void unlock() = 0;
